+++ src/Symfony/Component/HttpKernel/Fragment/InlineFragmentRenderer.php
@@ -124,8 +124,8 @@ class InlineFragmentRenderer extends RoutableFragmentRenderer
         if ($request->getDefaultLocale() !== $request->getLocale()) {
             $subRequest->setLocale($request->getLocale());
         }
-        if ($request->attributes->has('_stateless')) {
-            $subRequest->attributes->set('_stateless', $request->attributes->get('_stateless'));
+        if (null !== $request->isStateless()) {
+            $subRequest->setStateless($request->isStateless());
         }
         if ($request->attributes->has('_check_controller_is_allowed')) {
             $subRequest->attributes->set('_check_controller_is_allowed', $request->attributes->get('_check_controller_is_allowed'));